reprise des authorisations , de la mise en page et des migrations suite a un ajout qui a engendrer un grand nombre de bug . projet ok

+++ LITrevu/blog/views.py
@@ -4,14 +4,13 @@ from .models import Ticket, Review, UserFollows
 from .forms import TicketForm, ReviewForm, FollowUsersForm
 from itertools import chain
 from django.shortcuts import get_object_or_404
+from django.http import HttpResponseForbidden
 from operator import attrgetter
 
 
 @login_required
 def home(request):
     followed_users = request.user.following.values_list('followed_user', flat=True)
-
-
     tickets = Ticket.objects.filter(user__in=followed_users) | Ticket.objects.filter(user=request.user)
     reviews = Review.objects.filter(user__in=followed_users) | Review.objects.filter(user=request.user)
 
@@ -20,12 +19,18 @@ def home(request):
         key=lambda instance: instance.time_created,
         reverse=True
     )
+    # Ajouter un indicateur si la liste est vide
+    is_empty = not combined_list
 
     # Ajout d'une clé 'type' pour chaque élément dans la liste
     for item in combined_list:
         item.item_type = 'ticket' if isinstance(item, Ticket) else 'review'
     
-    return render(request, 'blog/flux.html', context={'combined_list': combined_list, 'followed_users': followed_users})
+    return render(request, 'blog/flux.html', context={
+        'combined_list': combined_list,
+        'followed_users': followed_users,
+        'is_empty': is_empty,
+        })
 
 @login_required
 @permission_required('blog.add_ticket', raise_exception=True)
@@ -46,12 +51,14 @@ def create_ticket(request):
 def update_ticket(request, ticket_id):  
     ticket = Ticket.objects.get(id=ticket_id)
     if ticket.user != request.user:
-        return redirect('home')
+        return redirect('permission_denied')
+    if ticket.user != request.user:
+        return redirect('flux')
     if request.method =='POST':
         form = TicketForm(request.POST, request.FILES, instance=ticket)
         if form.is_valid():
             form.save()
-            return redirect('home')
+            return redirect('flux')
     else:
         form = TicketForm(instance=ticket)
     return render ( request, 'blog/update_ticket.html', context={'form':form})
@@ -60,15 +67,21 @@ def update_ticket(request, ticket_id):
 @permission_required('blog.delete_ticket', raise_exception=True)
 def delete_ticket(request, ticket_id):
     ticket = Ticket.objects.get(id=ticket_id)
+    if ticket.user != request.user:
+        return redirect('permission_denied')
     if request.method =='POST':
         ticket.delete()
-        return redirect('home')
+        return redirect('flux')
     return render(request, 'blog/delete_ticket.html', context={'ticket':ticket})
 
 @login_required
 @permission_required('blog.add_review', raise_exception=True)
 def create_review(request, ticket_id):
     ticket = Ticket.objects.get(id=ticket_id)
+     # Vérifie si l'utilisateur a déjà publié une critique pour ce billet
+    if Review.objects.filter(ticket=ticket, user=request.user).exists():
+        return redirect('permission_denied')  
+    
     if request.method =='POST':
         form = ReviewForm(request.POST)
         if form.is_valid():
@@ -76,20 +89,24 @@ def create_review(request, ticket_id):
             review.ticket = ticket
             review.user = request.user
             review.save()
-            return redirect('home')
+            return redirect('flux')
     else:
         form = ReviewForm()
-    return render (request, 'blog/create_review.html', context={'form':form, 'ticket':ticket})
+    return render (request,
+                   'blog/create_review.html',
+                   context={'form':form, 'ticket':ticket})
 
 @login_required
 @permission_required('blog.change_review', raise_exception=True)
 def update_review(request, review_id):
     review = Review.objects.get(id=review_id)
+    if review.user != request.user:
+        return redirect('permission_denied')
     if request.method =='POST':
         form = ReviewForm(request.POST, request.FILES, instance=review)
         if form.is_valid():
             form.save()
-            return redirect('home')
+            return redirect('flux')
     else:
         form = ReviewForm(instance=review)
     return render ( request, 'blog/update_review.html', context={'form':form})
@@ -98,9 +115,11 @@ def update_review(request, review_id):
 @permission_required('blog.delete_review', raise_exception=True)
 def delete_review(request, review_id):
     review = Review.objects.get(id=review_id)
+    if review.user != request.user:
+        return redirect('permission_denied')
     if request.method =='POST':
         review.delete()
-        return redirect('home')
+        return redirect('flux')
     return render(request, 'blog/delete_review.html', context={'review':review})
 
 
@@ -124,7 +143,7 @@ def create_ticket_and_review(request):
             review.user = request.user
             review.save()
             
-            return redirect('home')
+            return redirect('flux')
     else:
         form_ticket = TicketForm()
         form_review = ReviewForm()
@@ -152,10 +171,16 @@ def user_posts(request):
         reverse=True
     )
 
+    # Ajouter un indicateur si la liste est vide
+    is_empty = not combined_list
+
     for item in combined_list:
         item.item_type = 'ticket' if isinstance(item, Ticket) else 'review'
 
-    return render(request, 'blog/user_posts.html', {'combined_list': combined_list})
+    return render(request, 'blog/user_posts.html', context={
+        'combined_list': combined_list,
+        'is_empty': is_empty,
+        })
 
 
 @login_required
@@ -188,3 +213,9 @@ def delete_followed_user(request, follow_id):
 
     # Rendre le template avec le nom de l'utilisateur suivi
     return render(request, 'blog/delete_followed_user.html', {'followed_user': follow_instance.followed_user})
+
+
+def permission_denied_view(request):
+    return render(request, 'blog/permission_denied.html', {
+        'message': "Vous n'avez pas les autorisations nécessaires pour accéder à cette ressource."
+    })
